+++ src/app/download-csv.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { DownloadCSVService } from './download-csv.service';
+
+describe('DownloadCSVService', () => {
+	beforeEach(() => TestBed.configureTestingModule({}));
+
+	it('should be created', () => {
+		const service: DownloadCSVService = TestBed.get(DownloadCSVService);
+		expect(service).toBeTruthy();
+	});
+});
